split components into different files to fix nextjs hot reload
