+++ apps/chat/src/components/Chat/Publish/PublicationHandler/PublicationReviewResources/view-props.ts
@@ -1,0 +1,5 @@
+import { PublicationResource } from '@/src/types/publication';
+
+export interface EntityPublicationResourcesProps {
+  resources: PublicationResource[];
+}
